24 Separate bounding_box module
