+++ bioio_czi/metadata_ome.py
@@ -1,6 +1,6 @@
 import os
 from pathlib import Path
-from typing import Dict, Tuple, Union
+from typing import Union
 from xml.etree import ElementTree as ET
 
 import lxml.etree
@@ -8,8 +8,6 @@ from bioio_base.types import PathLike
 from ome_types import OME
 
 Metadata = ET.Element
-# Example bounding box: {'X': (0, 100), 'Y': (0, 100), 'Z': (0, 20)}
-BoundingBox = Dict[str, Tuple[int, int]]
 
 
 class UnsupportedMetadataError(Exception):
